fix output alignment when printing marks
the `marks` function right-justifies mark names, to make them easier to
read. However, because of the escape sequences for colors in the output,
the printf command doesn't calculate the padding correctly, when
attempting to right-justify the mark names. This results in the mark
names not actually being right-justified correctly, with short names not
being padded enough.

this commit fixes this problem by only applying padding to the actual
mark name itself, and not the name+colors.